Movet settings to repo folder
Doing this even if we are planning to rename this folder, we need things
in the same place.
